+++ src/index.js
@@ -1,13 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Relay from 'react-relay';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 import { AppContainer } from 'react-hot-loader';
-import networkLayer from 'shared/utils/networkLayer';
 import Root from './Root';
 
 injectTapEventPlugin();
-Relay.injectNetworkLayer(networkLayer);
 
 const render = Component => {
   ReactDOM.render(